requirements.txt
PlexAPI~=4.12.0
python-dotenv~=0.20.0
SQLAlchemy
psutil~=5.9.4
requests~=2.31.0
tqdm~=4.64.1
numpy~=1.24.1
Pillow~=9.4.0
moviepy~=1.0.3
py-cpuinfo~=9.0.0